Refactor chart vars
